Conversion to non-static, multi-document handling ability
Reading about TF*IDF being required confused me since we are not given a set corpus to work on and TF*IDF doesn't make sense on a single document.
However it turns out the two given sites are the entire corpus, so I made the program work to create an object, PipelineObject, that contains all the data I have found or created for each document and compartmentalised it in an index-able object.